fix: linting issue and move stubs into own file
